break nailgun crate into multiple files
[ci skip-build-wheels]
