+++ src/rust/engine/nailgun/src/lib.rs
@@ -31,435 +31,9 @@
 #[cfg(test)]
 mod tests;
 
-use std::collections::HashMap;
-use std::io;
-use std::net::Ipv4Addr;
-use std::os::unix::io::{AsRawFd, FromRawFd, IntoRawFd, RawFd};
-use std::path::PathBuf;
-use std::sync::Arc;
-use tokio::io::AsyncWriteExt;
+mod client;
+mod server;
 
-use bytes::Bytes;
-use futures::channel::{mpsc, oneshot};
-use futures::{future, sink, stream, FutureExt, SinkExt, Stream, StreamExt, TryStreamExt};
-use log::{debug, error, info};
+pub use client::client_execute;
 pub use nails::execution::ExitCode;
-use nails::execution::{self, sink_for, stream_for, ChildInput, ChildOutput};
-use nails::{Child, Config, Nail};
-use tokio::fs::File;
-use tokio::net::TcpListener;
-use tokio::net::TcpStream;
-use tokio::sync::{Notify, RwLock};
-
-use task_executor::Executor;
-
-async fn handle_client_output(
-  mut stdio_read: impl Stream<Item = ChildOutput> + Unpin,
-) -> Result<(), io::Error> {
-  let mut stdout = tokio::io::stdout();
-  let mut stderr = tokio::io::stderr();
-  while let Some(output) = stdio_read.next().await {
-    match output {
-      ChildOutput::Stdout(bytes) => stdout.write_all(&bytes).await?,
-      ChildOutput::Stderr(bytes) => stderr.write_all(&bytes).await?,
-      ChildOutput::Exit(_) => {
-        // NB: We ignore exit here and allow the main thread to handle exiting.
-        break;
-      }
-    }
-  }
-  Ok(())
-}
-
-async fn handle_client_input(mut stdin_write: mpsc::Sender<ChildInput>) -> Result<(), io::Error> {
-  use nails::execution::send_to_io;
-  let mut stdin = stream_for(tokio::io::stdin());
-  while let Some(input_bytes) = stdin.next().await {
-    stdin_write
-      .send(ChildInput::Stdin(input_bytes?))
-      .await
-      .map_err(send_to_io)?;
-  }
-  stdin_write
-    .send(ChildInput::StdinEOF)
-    .await
-    .map_err(send_to_io)?;
-  Ok(())
-}
-
-async fn client_execute_helper(
-  port: u16,
-  command: String,
-  args: Vec<String>,
-  env: Vec<(String, String)>,
-) -> Result<i32, String> {
-  use nails::execution::{child_channel, Command};
-
-  let config = Config::default();
-  let command = Command {
-    command,
-    args,
-    env,
-    working_dir: std::path::PathBuf::from("/dev/null"),
-  };
-
-  let (stdio_write, stdio_read) = child_channel::<ChildOutput>();
-  let (stdin_write, stdin_read) = child_channel::<ChildInput>();
-
-  let output_handler = tokio::spawn(handle_client_output(stdio_read));
-  let _input_handler = tokio::spawn(handle_client_input(stdin_write));
-
-  let localhost = std::net::Ipv4Addr::new(127, 0, 0, 1);
-  let addr = (localhost, port);
-
-  let socket = TcpStream::connect(addr)
-    .await
-    .map_err(|err| format!("Nailgun client error connecting to localhost: {}", err))?;
-  let exit_code: ExitCode =
-    nails::client_handle_connection(config, socket, command, stdio_write, stdin_read)
-      .await
-      .map_err(|err| format!("Nailgun client error: {}", err))?;
-
-  let () = output_handler
-    .await
-    .map_err(|join_error| format!("Error joining nailgun client task: {}", join_error))?
-    .map_err(|err| format!("Nailgun client output error: {}", err))?;
-
-  Ok(exit_code.0)
-}
-
-pub async fn client_execute(
-  port: u16,
-  command: String,
-  args: Vec<String>,
-  env: Vec<(String, String)>,
-  exit_receiver: oneshot::Receiver<()>,
-) -> Result<i32, String> {
-  use future::Either;
-
-  let execution_future = client_execute_helper(port, command, args, env).boxed();
-
-  match future::select(execution_future, exit_receiver).await {
-    Either::Left((execution_result, _exit_receiver_fut)) => {
-      debug!("Nailgun client future finished");
-      execution_result
-    }
-    Either::Right((_exited, _execution_result_future)) => {
-      Err("Exiting nailgun client future via explicit quit message".to_string())
-    }
-  }
-}
-
-pub struct Server {
-  exit_sender: oneshot::Sender<()>,
-  exited_receiver: oneshot::Receiver<Result<(), String>>,
-  port: u16,
-}
-
-impl Server {
-  ///
-  /// Spawn the server on a background Task.
-  ///
-  /// The port provided here may be `0` in order to request a random port. A caller can use
-  /// `Server.port()` to determine what port was actually selected.
-  ///
-  pub async fn new(
-    executor: Executor,
-    port_requested: u16,
-    runner: impl RawFdRunner + 'static,
-  ) -> Result<Server, String> {
-    let listener = TcpListener::bind((Ipv4Addr::new(127, 0, 0, 1), port_requested))
-      .await
-      .map_err(|e| format!("Could not bind to port {}: {:?}", port_requested, e))?;
-    let port_actual = listener
-      .local_addr()
-      .map_err(|e| format!("No local address for listener: {:?}", e))?
-      .port();
-
-    // NB: The C client requires noisy_stdin (see the `nails` crate for more info), but neither
-    // `nails` nor the pants python client do.
-    let config = nails::Config::default().noisy_stdin(false);
-    let nail = RawFdNail {
-      executor: executor.clone(),
-      runner: Arc::new(runner),
-    };
-
-    let (exited_sender, exited_receiver) = oneshot::channel();
-    let (exit_sender, exit_receiver) = oneshot::channel();
-
-    let _join = executor.spawn(Self::serve(
-      executor.clone(),
-      config,
-      nail,
-      exit_receiver,
-      exited_sender,
-      listener,
-    ));
-
-    Ok(Server {
-      exit_sender,
-      exited_receiver,
-      port: port_actual,
-    })
-  }
-
-  ///
-  /// The main loop of the server. Public for testing.
-  ///
-  pub(crate) async fn serve(
-    executor: Executor,
-    config: nails::Config,
-    nail: impl Nail,
-    should_exit: oneshot::Receiver<()>,
-    exited: oneshot::Sender<Result<(), String>>,
-    listener: TcpListener,
-  ) {
-    let exit_result = Self::accept_loop(executor, config, nail, should_exit, listener).await;
-    info!("Server exiting with {:?}", exit_result);
-    let _ = exited.send(exit_result);
-  }
-
-  async fn accept_loop(
-    executor: Executor,
-    config: nails::Config,
-    nail: impl Nail,
-    mut should_exit: oneshot::Receiver<()>,
-    mut listener: TcpListener,
-  ) -> Result<(), String> {
-    // While connections are ongoing, they acquire `read`; before shutting down, the server
-    // acquires `write`.
-    let ongoing_connections = Arc::new(RwLock::new(()));
-
-    let result = loop {
-      let tcp_stream = match future::select(listener.accept().boxed(), should_exit).await {
-        future::Either::Left((Ok((tcp_stream, _addr)), s_e)) => {
-          // Got a connection.
-          should_exit = s_e;
-          tcp_stream
-        }
-        future::Either::Left((Err(e), _)) => {
-          break Err(format!("Server failed to accept connections: {}", e));
-        }
-        future::Either::Right((_, _)) => {
-          break Ok(());
-        }
-      };
-
-      debug!("Accepted connection: {:?}", tcp_stream);
-
-      // There is a slightly delicate dance here: we wait for a connection to have acquired the
-      // ongoing connections lock before proceeding to the next iteration of the loop. This
-      // prevents us from observing an empty lock and exiting before the connection has actually
-      // acquired it. Unfortunately we cannot acquire the lock in this thread and then send the
-      // guard to the other thread due to its lifetime bounds.
-      let connection_started = Arc::new(Notify::new());
-      let _join = executor.spawn({
-        let config = config.clone();
-        let nail = nail.clone();
-        let connection_started = connection_started.clone();
-        let ongoing_connections = ongoing_connections.clone();
-        async move {
-          let ongoing_connection_guard = ongoing_connections.read().await;
-          connection_started.notify();
-          let result = nails::server_handle_connection(config, nail, tcp_stream).await;
-          std::mem::drop(ongoing_connection_guard);
-          result
-        }
-      });
-      connection_started.notified().await;
-    };
-
-    // Before exiting, acquire write access on the ongoing_connections lock to prove that all
-    // connections have completed.
-    debug!("Server waiting for connections to complete...");
-    let _ = ongoing_connections.write().await;
-    debug!("All connections completed.");
-    result
-  }
-
-  ///
-  /// The port that the server is listening on.
-  ///
-  pub fn port(&self) -> u16 {
-    self.port
-  }
-
-  ///
-  /// Returns a Future that will shut down the server by:
-  /// 1. stopping accepting new connections
-  /// 2. waiting for all ongoing connections to have completed
-  ///
-  pub async fn shutdown(self) -> Result<(), String> {
-    // If we fail to send the exit signal, it's because the task is already shut down.
-    let _ = self.exit_sender.send(());
-    self
-      .exited_receiver
-      .await
-      .map_err(|_| "Server exited uncleanly.".to_owned())?
-  }
-}
-
-#[derive(Clone)]
-pub enum ServerState {
-  Bound(u16),
-  Exited(Result<(), String>),
-}
-
-pub struct RawFdExecution {
-  pub cmd: execution::Command,
-  pub stdin_fd: RawFd,
-  pub stdout_fd: RawFd,
-  pub stderr_fd: RawFd,
-}
-
-///
-/// Implementations of this trait should _not_ close the input file handles, and should let the
-/// caller do so instead.
-///
-pub trait RawFdRunner: Fn(RawFdExecution) -> ExitCode + Send + Sync {}
-
-impl<T: Fn(RawFdExecution) -> ExitCode + Send + Sync> RawFdRunner for T {}
-
-///
-/// A Nail implementation that proxies stdio to file handles that can be consumed by the given
-/// callback function.
-///
-/// If any of stdio is a tty (detected via environment variables special cased in our nailgun
-/// client), we ignore the input fds and open new fds directly to the tty path, which happens to
-/// be addressable as a file in OSX and Linux. In that case, there is no middle-man, and we
-/// ignore the fds opened by the nailgun server for data sent via the protocol.
-///
-#[derive(Clone)]
-struct RawFdNail {
-  executor: Executor,
-  runner: Arc<dyn RawFdRunner>,
-}
-
-impl Nail for RawFdNail {
-  fn spawn(
-    &self,
-    cmd: execution::Command,
-    input_stream: mpsc::Receiver<ChildInput>,
-  ) -> Result<Child, io::Error> {
-    let env = cmd.env.iter().cloned().collect::<HashMap<_, _>>();
-
-    // Handle stdin.
-    let (stdin_handle, stdin_sink) = Self::input(Self::ttypath_from_env(&env, 0))?;
-    let accepts_stdin = if let Some(mut stdin_sink) = stdin_sink {
-      let mut input_stream = input_stream.filter_map(|child_input| {
-        Box::pin(async move {
-          match child_input {
-            ChildInput::Stdin(bytes) => Some(Ok(bytes)),
-            ChildInput::StdinEOF => None,
-          }
-        })
-      });
-      let _join = self.executor.spawn(async move {
-        stdin_sink.send_all(&mut input_stream).map(|_| ()).await;
-      });
-      true
-    } else {
-      // Stdin will be handled directly by the TTY.
-      false
-    };
-
-    // And stdout/stderr.
-    let (stdout_stream, stdout_handle) = Self::output(Self::ttypath_from_env(&env, 1))?;
-    let (stderr_stream, stderr_handle) = Self::output(Self::ttypath_from_env(&env, 2))?;
-
-    // Spawn the underlying function as a blocking task, and capture its exit code to append to the
-    // output stream.
-    let nail = self.clone();
-    let exit_code_future = self.executor.spawn_blocking(move || {
-      // NB: This closure captures the stdio handles, and will drop/close them when it completes.
-      (nail.runner)(RawFdExecution {
-        cmd,
-        stdin_fd: stdin_handle.as_raw_fd(),
-        stdout_fd: stdout_handle.as_raw_fd(),
-        stderr_fd: stderr_handle.as_raw_fd(),
-      })
-    });
-
-    // Fully consume the stdout/stderr streams before waiting on the exit stream.
-    let stdout_stream = stdout_stream.map_ok(ChildOutput::Stdout);
-    let stderr_stream = stderr_stream.map_ok(ChildOutput::Stderr);
-    let exit_stream = exit_code_future
-      .into_stream()
-      .map(|exit_code| Ok(ChildOutput::Exit(exit_code)));
-    let output_stream = stream::select(stdout_stream, stderr_stream)
-      .chain(exit_stream)
-      .map(|res| match res {
-        Ok(o) => o,
-        Err(e) => {
-          error!("IO error interacting with the runner: {:?}", e);
-          ChildOutput::Exit(ExitCode(-1))
-        }
-      })
-      .boxed();
-
-    Ok(Child {
-      output_stream,
-      accepts_stdin,
-    })
-  }
-}
-
-impl RawFdNail {
-  ///
-  /// Returns a tuple of a readable file handle and an optional sink for nails to send stdin to.
-  ///
-  /// In the case of a TTY, the file handle will point directly to the TTY, and no stdin data will
-  /// flow over the protocol. Otherwise, it will be backed by a new anonymous pipe, and data should
-  /// be copied to the returned Sink.
-  ///
-  fn input(
-    tty_path: Option<PathBuf>,
-  ) -> Result<(Box<dyn AsRawFd + Send>, Option<impl sink::Sink<Bytes>>), io::Error> {
-    if let Some(tty_path) = tty_path {
-      Ok((Box::new(std::fs::File::open(tty_path)?), None))
-    } else {
-      let (stdin_reader, stdin_writer) = os_pipe::pipe()?;
-      let write_handle =
-        File::from_std(unsafe { std::fs::File::from_raw_fd(stdin_writer.into_raw_fd()) });
-      Ok((Box::new(stdin_reader), Some(sink_for(write_handle))))
-    }
-  }
-
-  ///
-  /// Returns a tuple of a possibly empty Stream for nails to read data from, and a writable file handle.
-  ///
-  /// See `Self::input` and the struct's rustdoc for more info on the TTY case.
-  ///
-  #[allow(clippy::type_complexity)]
-  fn output(
-    tty_path: Option<PathBuf>,
-  ) -> Result<
-    (
-      stream::BoxStream<'static, Result<Bytes, io::Error>>,
-      Box<dyn AsRawFd + Send>,
-    ),
-    io::Error,
-  > {
-    if let Some(tty_path) = tty_path {
-      let tty = std::fs::OpenOptions::new()
-        .write(true)
-        .create(false)
-        .open(tty_path)?;
-      Ok((stream::empty().boxed(), Box::new(tty)))
-    } else {
-      let (stdin_reader, stdin_writer) = os_pipe::pipe()?;
-      let read_handle =
-        File::from_std(unsafe { std::fs::File::from_raw_fd(stdin_reader.into_raw_fd()) });
-      Ok((stream_for(read_handle).boxed(), Box::new(stdin_writer)))
-    }
-  }
-
-  ///
-  /// Corresponds to `ttynames_to_env` in `nailgun_protocol.py`. See this struct's rustdocs.
-  ///
-  fn ttypath_from_env(env: &HashMap<String, String>, fd_number: usize) -> Option<PathBuf> {
-    env
-      .get(&format!("NAILGUN_TTY_PATH_{}", fd_number))
-      .map(PathBuf::from)
-  }
-}
+pub use server::{RawFdExecution, Server};
